split parser impl in two files
